pyproject.toml
[project]
name = "rize-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "eren-olympic",email = "eren.olympic.14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-frontmatter (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
